+++ Sources/StackKit/Divider/DividerResultBuilder.swift
