+++ Sources/StackKit/Divider/DividierView.swift
@@ -1,0 +1,16 @@
+import UIKit
+
+class DividerView: UIView, _Divider {
+    
+    var maxLength: CGFloat = .greatestFiniteMagnitude
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        isUserInteractionEnabled = false
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
