+++ Sources/StackKit/Spacer/SpacerResultBuilder.swift
