+++ Sources/StackKit/Spacer/_Spacer.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+public protocol _Spacer {
+    var length: CGFloat { get set }
+    var min: CGFloat { get set }
+    var max: CGFloat { get set }
+}
